requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0
google-generativeai>=0.3.0
google-ads>=22.0.0
google-analytics-data>=0.16.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
